Fix inconsistency with update_fields not containing the version field
Previously, if a model was saved with update_fields, the version on the
instance would be inconsistent with the version in the database. Now,
they match so that the model can be saved again.

Fixes #6, but not but implicitly adding version to update_fields,
instead allowing the use of update_fields to bypass versioning.